requirements.txt
langgraph==1.0.1
langchain==1.0.1
langchain-openai==1.0.0
llama-index-core==0.14.5
llama-index-llms-openai==0.6.5